refactor(pr-check): reduce duplication in oneOf/anyOf tests
