+++ packages/schema/test/types.ts
@@ -1,5 +1,5 @@
 import { type Schema } from '../src/schema';
-import { bigint, boolean, object, oneOf, optional } from '../src';
+import { anyOf, bigint, boolean, object, oneOf, optional } from '../src';
 import { string } from '../src/types/string';
 import { number } from '../src/types/number';
 import { validateAndMap } from '../src/schema';
@@ -73,3 +73,24 @@ export enum Color {
   Green = 'green',
   Blue = 'blue',
 }
+
+const schema1 = object({
+  type: ['type', string()],
+  name: ['name', string()],
+  age: ['age', number()],
+});
+
+const schema2 = object({
+  type: ['type', string()],
+  title: ['title', string()],
+  rating: ['rating', string()],
+});
+
+const discriminatorMap = {
+  object1: schema1,
+  object2: schema2,
+};
+
+
+export const anyOfWithDiscriminator = anyOf([schema1, schema2], discriminatorMap, 'type');
+export const oneOfWithDiscriminator = oneOf([schema1, schema2], discriminatorMap, 'type');
