+++ packages/schema/test/types/oneOf.test.ts
@@ -16,6 +16,7 @@ import {
   Color,
   Human,
   humanSchema,
+  oneOfWithDiscriminator,
   Person,
   personSchema,
 } from '../types';
@@ -253,31 +254,13 @@ describe('OnyOf', () => {
     });
 
     it('should map oneOf with discriminator', () => {
-      const schema1 = object({
-        type: ['type', string()],
-        name: ['name', string()],
-        age: ['age', number()],
-      });
-
-      const schema2 = object({
-        type: ['type', string()],
-        title: ['title', string()],
-        rating: ['rating', string()],
-      });
-
-      const discriminatorMap = {
-        object1: schema1,
-        object2: schema2,
-      };
-
       const input = {
         type: 'object1', // The discriminator field value that matches schema1
         name: 'John',
         age: 30,
       };
 
-      const schema = oneOf([schema1, schema2], discriminatorMap, 'type');
-      const output = validateAndMap(input, schema);
+      const output = validateAndMap(input, oneOfWithDiscriminator);
 
       expect(output.errors).toBeFalsy();
       expect((output as any).result).toStrictEqual(input); // The input should be unchanged since it matches schema1
@@ -459,31 +442,13 @@ describe('OnyOf', () => {
     });
 
     it('should unmap oneOf with discriminator', () => {
-      const schema1 = object({
-        type: ['type', string()],
-        name: ['name', string()],
-        age: ['age', number()],
-      });
-
-      const schema2 = object({
-        type: ['type', string()],
-        title: ['title', string()],
-        rating: ['rating', string()],
-      });
-
-      const discriminatorMap = {
-        object1: schema1,
-        object2: schema2,
-      };
-
       const input = {
         type: 'object1', // The discriminator field value that matches schema1
         name: 'John',
         age: 30,
       };
 
-      const schema = oneOf([schema1, schema2], discriminatorMap, 'type');
-      const output = validateAndUnmap(input, schema);
+      const output = validateAndUnmap(input, oneOfWithDiscriminator);
 
       expect(output.errors).toBeFalsy();
       expect((output as any).result).toStrictEqual(input); // The input should be unchanged since it matches schema1
